Almost finale version

+++ cloudServer/execution_service/function_execution.go
@@ -2,19 +2,22 @@ package execution_service
 
 import (
 	"bytes"
+	"edgeServer/function_handling"
 	"edgeServer/storage_service"
 	"log"
 	"os"
 	"os/exec"
 	"strconv"
+	"strings"
 )
 
 func ExecuteAndDetachFunctionWasmer(functionName string, parameter string) string {
 	cmd := exec.Command("")
+	fNameExecutable := strings.Split(functionName, "-")[0]
 	if len(parameter) > 0 {
-		cmd = exec.Command("wasmer", "run", "./wasm_files/"+functionName+".wasm", "--", "-e", parameter)
+		cmd = exec.Command("wasmer", "run", "./wasm_files/"+fNameExecutable+".wasm", "--", "-e", parameter)
 	} else {
-		cmd = exec.Command("wasmer", "run", "./wasm_files/"+functionName+".wasm")
+		cmd = exec.Command("wasmer", "run", "./wasm_files/"+fNameExecutable+".wasm")
 	}
 	var outb, errb bytes.Buffer
 	cmd.Stdout = &outb
@@ -22,7 +25,7 @@ func ExecuteAndDetachFunctionWasmer(functionName string, parameter string) strin
 	_ = cmd.Start()
 	log.Println(outb.String())
 	storage_service.SetValue(functionName, strconv.Itoa(cmd.Process.Pid))
-	go func() {StopFunction(functionName, strconv.Itoa(cmd.Process.Pid)) }()
+	//go func() {StopFunction(functionName, strconv.Itoa(cmd.Process.Pid)) }()
 	_ = cmd.Wait()
 	log.Println("deleting!")
 	storage_service.DeleteKey(functionName)
@@ -33,15 +36,15 @@ func StopFunction(functionName string, pid string) {
 	nPid, _ := strconv.Atoi(pid)
 	log.Printf("finding %d", nPid)
 	proc, _ := os.FindProcess(nPid)
-	//if err != nil {
-	//	log.Printf("Could not find process %s with pid %s %v", functionName, pid, err)
-	//	return
-	//}
-	if proc != nil {
+	ans, err := function_handling.IsProcessActiveUsingPid(pid);
+	if err != nil {
+		log.Printf("Could not find process %s with pid %s %v", functionName, pid, err)
+		return
+	}
+	if proc != nil && ans {
 		_ = proc.Kill()
 		storage_service.DeleteKey(functionName)
 		return
 	}
 	log.Printf("Could not find process %s with pid %s", functionName, pid)
-	log.Println("ended delete")
 }
